BUG#: 8692 TITLE: valgrind error reported in AnonymousPipe::writeBuffer. DESCRIPTION: Issue with structure padding in SCMOResolutionTable.

+++ src/Pegasus/Common/SCMOStreamer.h
@@ -47,7 +47,7 @@ struct SCMOResolutionTable
     // Though we really store a pointer here, it is stored as Uint64 to
     // become independent from 64bit versus 32bit incarnations of the struct.
     Uint64 scmbptr;
-    Uint32 index;
+    Uint64 index;
 };
 
 #define PEGASUS_ARRAY_T SCMOResolutionTable
